service/birds_detection/model-serving/requirements.txt
common-code[test] @ git+https://github.com/swiss-ai-center/common-code.git@main
numpy
tensorflow
scipy
pillow
librosa